Bill S5183 version A - Robert Ortt
BILL NUMBER: S5183A SPONSOR: ORTT TITLE OF BILL: An act to amend the tax law, in relation to extending the period during which the county of Orleans is authorized to impose additional rates of sales and compensating use taxes PURPOSE: This legislation would extend the sales and compensating use tax in the county of Orleans. SUMMARY OF PROVISIONS: Section 1. the sales and compensating tax rate, which is an additional 1% to the 3% rate authorized, is extended to 2027.