Bill S5928 version original - George Borrello
BILL NUMBER: S5928 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the penal law, in relation to harassment in the first degree PURPOSE OR GENERAL IDEA OF BILL: This bill will make it a class B misdemeanor (harassment in the first degree) when someone strikes, shoves, kicks or otherwise subjects a school employee (i.e. teacher) to physical contact or attempts the same, thereby making it possible to prosecute such person who is under the age of 16 as a juvenile delinquent.