Bill S5922 version original - George Borrello
BILL NUMBER: S5922 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the tax law, in relation to exemptions to the petroleum business tax PURPOSE: This legislation would exclude Diesel and Motor fuel used in Cranes that are used off-road from the Petroleum Business Tax. SUMMARY OF PROVISIONS: Amends subdivision (a) of section 301-b of the tax law by adding a new paragraph (9) to include Diesel and Motor fuel used in Cranes.