Bill S5952 version original - Timothy M. Kennedy
BILL NUMBER: S5952 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to providing for the electronic transmission of certain documents PURPOSE OR GENERAL IDEA OF BILL: To allow for the electronic transmission of vehicle titles from the Department of Motor Vehicles to the vehicle owner. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends Vehicle and Traffic Law Article 46 Section 2101 to include electronic transmission.