Bill S5943 version original - James Skoufis
BILL NUMBER: S5943 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the election law, in relation to the order in which candidates appear on the ballot PURPOSE: To increase voter participation by creating a standard order in which candidates appear on the ballot. SUMMARY OF PROVISIONS: Section 1 amends the election law section 7-104 to require the executive candidates to be listed first, such as the President of the United States and Governor. Followed by other federal offices and state offices.