Bill S5938 version original - Jake Ashby
BILL NUMBER: S5938 SPONSOR: ASHBY TITLE OF BILL: An act to amend the civil service law, in relation to allowing reser- vists who have served or who are currently serving to be eligible for two and one-half points on civil service examinations PURPOSE: Allows reservists who have served or who are currently serving in the United States armed forces, to be eligible for 2.5 points on civil service examinations. SUMMARY OF PROVISIONS: Section 1.