Bill S5950 version original - James Skoufis
BILL NUMBER: S5950 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the penal law, in relation to aggravated harassment in the first degree PURPOSE: The purpose of this legislation is to include the display of a swastika on a billboard, aerial display or by a similar method for uninvited public viewing. SUMMARY OF PROVISIONS: This legislation will amend Section 240.