Bill S4742A version A - Jeremy Cooney
BILL NUMBER: S4742A SPONSOR: COONEY TITLE OF BILL: An act to amend the alcoholic beverage control law, in relation to authorizing special permits to remain open during certain hours of the morning PURPOSE OR GENERAL IDEA OF BILL: This bill would allow restaurants and bars to apply for a special permit to stay open on January 1 between 4 am and 10 am. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Allows licensees to apply for a special permit to stay open on the morning of January 1 from 4 a.m.