Bill S5935 version original - Cordell Cleare
BILL NUMBER: S5935 SPONSOR: CLEARE TITLE OF BILL: An act to amend the judiciary law, in relation to enacting the "hope card act" PURPOSE: This bill would require that the Office of Court Administration develop a program for the issuance of "Hope Cards" to petitioners that have been granted a final order of protection. Hope Cards will function as a wallet-sized, laminated card that petitioners can carry with them, and hold the same validity as a traditional paper order of protection.