Bill S5978 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S5978 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend the tax law, in relation to establishing a tax credit for the adoption of a dog or cat PURPOSE: To grant an individual taxpayer a credit of one hundred dollars for each dog or cat adopted from an animal shelter, for a maximum of three dogs or cats per taxable year. SUMMARY OF PROVISIONS: Section 1: Allows a credit for the adoption of a maximum of three dogs or cats per taxable year from a qualifying adoption entity.