Bill S5979 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S5979 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend the real property actions and proceedings law and the penal law, in relation to clarifying the definition of "tenant" to exclude squatters PURPOSE: This legislation would clarify the definition of tenant to exclude squatters. SUMMARY OF PROVISIONS: Section 711 of the real property actions and proceedings law is amended to clarify the definition of tenant to exclude squatters.