Bill S2781A version A - Dan Stec
BILL NUMBER: S2781A SPONSOR: STEC TITLE OF BILL: An act to amend chapter 368 of the laws of 2008, amending the tax law relating to authorizing the county of Warren to impose an additional mortgage recording tax, in relation to extending the effectiveness ther- eof PURPOSE: To extend the state's authorization permitting the county of Warren to impose an additional mortgage tax through Dec. 1, 2025 SUMMARY OF PROVISIONS: Section 1 - Authorizes to impose an additional mortgage tax through Dec.