Bill S5982A version original - Monica Martinez
BILL NUMBER: S5982REVISED 4/4/23 SPONSOR: MARTINEZ TITLE OF BILL: An act to authorize the study of aviation ownership and flight oper- ations by the department of transportation PURPOSE: To direct the department of transportation to conduct a study of Aviation flight operations by residents of the state SUMMARY OF PROVISIONS: Section 1. authorizes and directs the department of transportation to conduct a study of general aviation flight operations.