Bill S5987 version original - Mario Mattera
BILL NUMBER: S5987 SPONSOR: MATTERA TITLE OF BILL: An act to amend the penal law, in relation to exempting retired peace officers from the crime of criminal possession of a firearm, rifle or shotgun in a sensitive location PURPOSE: To add retired peace officers to the exempt list of those who can continue to carry their legal firearms in sensitive locations. SUMMARY: Section 1 subdivision 3 of section 265.