Bill S6001 version original - Andrew J. Lanza
BILL NUMBER: S6001 SPONSOR: LANZA TITLE OF BILL: An act to amend the correction law, in relation to the confinement of incarcerated individuals under the age of eighteen PURPOSE: This bill prohibits the placement of inmates under the age of eighteen in solitary confinement. SUMMARY OF PROVISIONS: Section 1 amends Correction Law 137 (6). Section 2 provides an effective date. JUSTIFICATION: New York is one of two states that continue to sentence juveniles under the age of 18 as adults.