Bill S6033 version original - Kevin S. Parker
BILL NUMBER: S6033 SPONSOR: PARKER TITLE OF BILL: An act to amend the insurance law, in relation to requiring owners of firearms to obtain liability insurance PURPOSE: To amend the insurance law, to require firearm owners, prior to such ownership, to obtain and continuously maintain a policy of liability insurance to cover any damages resulting from the use of such firearm. SUMMARY OF PROVISIONS: Section 1: The insurance law is amended by adding a new section 2354.