Bill S6046 version original - Gustavo Rivera
BILL NUMBER: S6046 Revised 2/13/24 SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law, in relation to payment of claims submitted by healthcare agencies PURPOSE: Provides for payment of claims submitted by healthcare agencies SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds new subdivisions to section 3621 of the public health law as it relates to the payment of claims. This section provides language regarding the payor's responsibility for the payment of claims.