Bill S6051 version original - Dean Murray
BILL NUMBER: S6051 SPONSOR: MURRAY TITLE OF BILL: An act to amend the vehicle and traffic law and the criminal procedure law, in relation to increasing penalties for leaving the scene of an accident without reporting PURPOSE: To enact "Nick's Law" in relation to increasing penalties and fines for leaving the scene of an accident involving a death. SUMMARY OF PROVISIONS: Section 1 provides that the title of this act shall be known as "Nick's Law.