Bill S6024 version original - Joseph A. Griffo
BILL NUMBER: S6024 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the tax law, in relation to extending the authority of the county of Madison to impose an additional rate of sales and compen- sating use taxes PURPOSE: Extends the additional one percent sales and compensating use tax for Madison County. SUMMARY OF PROVISIONS: Section one amends clause 35 of Section 1210 of the tax law to extend the authority of Madison County to impose an additional one percent sales and compensating use tax.