Bill S6019 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S6019 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend the penal law, in relation to the criminal sale of a controlled substance to a person less than fourteen years of age PURPOSE: To amend the Penal Law to provide that the sale of controlled substances to minors under 14 may be prosecuted as a more serious offense. SUMMARY OF PROVISIONS: Adds § 220.