Bill S6045A version original - Jamaal Bailey
BILL NUMBER: S6045 SPONSOR: BAILEY TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to establishing a fetal infant mortality review board PURPOSE OR GENERAL IDEA: The purpose of this bill is to establish a fetal infant mortality review board in the city of New York. SUMMARY OF SPECIFIC PROVISIONS: Section one of the bill amends the relevant provision of the administra- tive code of the city of New York to create the fetal infant mortality review board.