Bill S6017 version original - Andrew J. Lanza
BILL NUMBER: S6017 SPONSOR: LANZA TITLE OF BILL: An act to amend the real property tax law, in relation to the veterans' real property tax exemption with respect to members of the reserves PURPOSE: Relates to eligibility for a real property tax exemption for members of the reserve. SUMMARY OF PROVISIONS: Section one amends subdivision 10 of section 458-a of the real property tax law, as amended by section 83 of part PP of chapter 56 of the laws of 2022. Section two is the effective date.