Bill S6042 version original - Kevin S. Parker
BILL NUMBER: S6042 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to the establishment of community housing waiting lists for adults within the office of mental health service system PURPOSE: This legislation would enable the state to track the wait time for persons with Psychiatric disabilities seeking supportive, supervised or congregate housing in the office of mental health system. SUMMARY OF PROVISIONS: Section 1: Section 7.