Bill S5996 version original - Andrew J. Lanza
BILL NUMBER: S5996 SPONSOR: LANZA TITLE OF BILL: An act to amend the tax law, in relation to increasing the maximum pension and annuity exclusion from federal adjusted gross income PURPOSE: This bill would increase the maximum pension and annuity exclusion from federal adjusted income to $22,000. SUMMARY OF PROVISIONS: Section one amends paragraph 3-a of subsection (c) of section 612 of the tax law, as amended by section 3 part I of chapter 59 of the laws of 2015.