Bill S6022 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S6022 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the general municipal law, in relation to the conducting of games of chance and bingo by certain organizations PURPOSE: The purpose of this bill is two-fold. First, to update the provisions that, governs the conduct of games of chance in New York. Article 9-A of the General Municipal Law was first codified in 1967 and has not been substantially modernized since that date.