Bill S6009A version original - Kevin S. Parker
BILL NUMBER: S6009 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to creating a chemi- cal dependence treatment bill of rights PURPOSE OR GENERAL IDEA OF BILL: To ensure that patients of chemical dependence treatment facilities are accorded appropriate rights. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Subdivision (c) of section 22.93 if the mental hygiene law, as added by chapter 558 of the laws of 1999, is amended and a new subdi- vision (d) is added.