Bill S6058 version original - Peter Oberacker
BILL NUMBER: S6058 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the tax law, in relation to providing for a credit against personal income tax for volunteer firefighters and members of volunteer ambulance corps PURPOSE: Gives state income tax credit to volunteer firefighters and members of a volunteer ambulance corps in good standing up to $2500. SUMMARY OF PROVISIONS: Section 1.