Bill S6067 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S6067 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend the criminal procedure law, in relation to factors for the issuance of an order of recognizance or bail PURPOSE OR GENERAL IDEA OF BILL: The bill will help protect more victims of domestic violence by expand- ing criminal procedure law section 510.30. The bill will require the court, when determining recognizance or bail in cases of domestic violence, to consider certain additional enumerated factors.