Bill S6084B version original - Nathalia Fernandez
BILL NUMBER: S6084 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the general business law, in relation to prohibiting the sale of xylazine PURPOSE OR GENERAL IDEA OF BILL: To prohibit the sale of xylazine without proof of its intended use for institutional or scientific purposes. SUMMARY OF PROVISIONS: Section 1 prohibits the sale and delivery of xylazine in the State with- out proof of its intended use for institutional or scientific purposes. Section 2 sets the effective date.