Bill S6092 version original - Dean Murray
BILL NUMBER: S6092 SPONSOR: MURRAY TITLE OF BILL: An act to amend the correction law, in relation to facilitating compli- ance with the sex offender registry for offenders without a registerable residence PURPOSE: To enhance laws relating to the registration of homeless sex offenders. SUMMARY OF PROVISIONS: Section 1 amends section 168-a of the corrections law to define "home- less sex offender" as a person required to register as a sex offender who does not have a registerable residence.