Bill S6064 version original - Robert Ortt
BILL NUMBER: S6064 SPONSOR: ORTT TITLE OF BILL: An act to amend the tax law, in relation to authorizing the county of Niagara to continue to impose an additional rate of sales and compensat- ing use taxes PURPOSE: This legislation would extend the sales and compensating use tax in the county of Niagara. SUMMARY OF PROVISIONS: Section 1. the sales and compensating tax rate, which is an additional 1% to the 3% rate authorized, is extended to 2025.