Bill S1677A version A - Joseph P. Addabbo Jr.
BILL NUMBER: S1677A SPONSOR: ADDABBO TITLE OF BILL: An act to amend the agriculture and markets law, in relation to increas- ing fines for selling disabled Equidae PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to update the fines for selling disa- bled horses at auctions. The bill also expands the protect to mules and donkeys. SUMMARY OF PROVISIONS: Section 1: Section 358 of the agriculture and markets law is amended. Section 2: Established the effect date.