Bill S6062 version original - Pete Harckham
BILL NUMBER: S6062 SPONSOR: HARCKHAM PURPOSE: Waives the residency requirement for the commissioner of planning, development, and public transportation for Putnam county SUMMARY OF PROVISIONS: Section 1 amends section 3 of public officers law to exempt the county of Putnam from the requirement mandating the commissioner of planning, development, and public transportation to reside in Putnam County. Section 2 states the effective date.