Bill S6090 version original - Dean Murray
BILL NUMBER: S6090 SPONSOR: MURRAY TITLE OF BILL: An act to amend the general business law, in relation to requiring fair business practices in franchises PURPOSE: This bill is to ensure continued fair practices and good faith dealings between a franchise and a franchisee. SUMMARY OF PROVISIONS: Section 1. The general business law is amended by adding a new article, 33-C. JUSTIFICATION: The relationship between a franchise and a franchisee is governed by their franchise agreements.