Bill S6112A version original - Toby Ann Stavisky
BILL NUMBER: S6112 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to the provision of applications for registration to licensees by the state education department SUMMARY OF SPECIFIC PROVISIONS: Section 1. Subdivision 2 of section 6502 of the education law, as amended by chapter 81 of the laws of 1995, is amended to read as follows: 2.