Bill S5611A version A - Thomas F. O'Mara
BILL NUMBER: S5611A SPONSOR: O'MARA TITLE OF BILL: An act to amend the tax law, in relation to extending the authorization of the county of Schuyler to impose an additional one percent of sales and compensating use taxes PURPOSE: To extend statutory authorization to Schuyler County to extend an addi- tional one percent sales and compensating use tax. SUMMARY OF PROVISIONS: Section 1 amends Clause 22 of subparagraph (i) of the opening paragraph of section 1210 of the tax law.