Bill S6108 version original - Dean Murray
BILL NUMBER: S6108 SPONSOR: MURRAY TITLE OF BILL: An act to amend the labor law, in relation to the definition of manual worker PURPOSE: To update the definition of manual worker in labor law. SUMMARY OF PROVISIONS: Section 1 of the bill amends the definition of manual worker in statute from a "mechanic, workingman or laborer" to a "worker whose primary duty is to engage in physical labor.