Bill S6103 version original - Liz Krueger
BILL NUMBER: S6103 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the public health law, in relation to the administration of vaccines to children who are at least fourteen years of age PURPOSE OR GENERAL IDEA OF BILL: To allow minors who are at least fourteen years of age to have certain vaccinations administered to him or herself. SUMMARY OF PROVISIONS: Section 1: Amends the public health law by adding a new section 2167-a.