Bill S6119 version original - Steve Rhoads
BILL NUMBER: S6119 SPONSOR: RHOADS TITLE OF BILL: An act to amend the public health law, in relation to the administration of the novel coronavirus/COVID-19 immunization PURPOSE OR GENERAL IDEA OF BILL: Relates to the administration of the novel coronavirus/COVID-19 immuni- zation. SUMMARY OF PROVISIONS: Section 1. The public health law is amended by adding a new section 2184 to read as follows; § 2184. Novel coronavirus/COVID-19; immunization. 1.