Bill S6140 version original - Cordell Cleare
BILL NUMBER: S6140 SPONSOR: CLEARE TITLE OF BILL: An act to amend the executive law, in relation to enacting the "Wander- ing Officers Act" prohibiting the hiring of certain persons as a police officer PURPOSE OR GENERAL IDEA OF BILL: Prevents people who have been fired as or who were in jeopardy of being fired as a police officer in any other jurisdiction from being hired as police officers in New York State. SUMMARY OF PROVISIONS: Section 1. Names the act as the Wandering Officers Act.