Bill S6121 version original - Monica Martinez
BILL NUMBER: S6121 SPONSOR: MARTINEZ TITLE OF BILL: An act directing the commissioner of taxation and finance to grant an automatic extension until May 15, 2023 for counties affected by a severe winter storm in December 2022 PURPOSE: The purpose of this bill is to grant an automatic extension until May 15, 2023 for counties affected by a severe winter storm in December 2022. SUMMARY OF PROVISIONS: Section 1.