Bill S6123 version original - Gustavo Rivera
BILL NUMBER: S6123 SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law, in relation to the submission of claims for services provided by home care agencies PURPOSE OR GENERAL IDEA OF BILL: To improve the process for claiming and billing for home and community- based providers. SUMMARY OF SPECIFIC PROVISIONS: The bill amends Public Health Law Article 36, to provide that home and community providers shall use federally approved forms for the claiming to insurance companies.