Bill S6118A version original - Brad Hoylman-Sigal
BILL NUMBER: S6118 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the insurance law, in relation to the definition of infertility and health insurance coverage for in-vitro fertilization PURPOSE: To ensure that all those wish to become parents through in-vitro ferti- lization (IVF) may have the coverage to do so by requiring insurance coverage for donor cycle. SUMMARY OF SPECIFIC PROVISIONS: Section 1 sets out the title of this act.