Bill S6129 version original - Kevin S. Parker
BILL NUMBER: S6129 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to establishing an appeals process for students denied the state resident tuition rate at a public college or university PURPOSE OR GENERAL IDEA OF BILL: To establish an appeals process for students who are denied the state resident tuition rate at any public university or college. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Section 355 of the education law is amended by adding a new subdivision 22.