Bill S6134 version original - Kevin S. Parker
BILL NUMBER: S6134 SPONSOR: PARKER TITLE OF BILL: An act to establish a task force to study the impact of a state-issued cryptocurrency on the state of New York PURPOSE OR GENERAL IDEA OF BILL: Establishes a task force to study the issuance of a state-issued crypto- currency in to raise revenue New York State. (NYcoin) SUMMARY OF PROVISIONS: Section 1: A task force is hereby established to study the impact of a state-issued cryptocurrency on the state of New York.