Bill S6122 version original - Monica Martinez
BILL NUMBER: S6122 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend chapter 130 of the laws of 1998, amending the general municipal law relating to temporary investments by local governments, in relation to extending the expiration of the provisions thereof PURPOSE: The purpose of this bill is to extend the authority granted to local governments to invest in additional types of products. SUMMARY OF PROVISIONS: Section 1.