Bill S6139 version original - Steve Rhoads
BILL NUMBER: S6139 SPONSOR: RHOADS TITLE OF BILL: An act to amend the election law, the administrative code of the city of New York, the public health law and the vehicle and traffic law, in relation to requiring a government issued photo identification card to be presented when casting a ballot PURPOSE: Requires a government photo identification card to be presented when casting a ballot. SUMMARY OF PROVISIONS: Section 1.