Bill S6120 version original - Monica Martinez
BILL NUMBER: S6120 Revised 4/4/23 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the education law, in relation to establishing a comput- er science graduation requirement; and providing for the repeal of certain provisions of such law relating thereto PURPOSE: The purpose of this bill is to make computer science a required course option for graduation from high school. SUMMARY OF PROVISIONS: Section 1.