Bill S6170 version original - Nathalia Fernandez
BILL NUMBER: S6170 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the education law, in relation to the licensing of sign language interpreters and creating a state board for sign language interpreting PURPOSE: The purpose of this bill is to create a licensing requirement for sign language interpreters. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill amends the education law by adding a new article 169 entitled Sign Language Interpreters to: 1.