Bill S6151A version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S6151 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend chapter 676 of the laws of 1978, amending the town law relating to payment in lieu of taxes for certain property in the town of Hempstead, in relation to extending the expiration thereof PURPOSE: Extending a PILOT program for Lido Beach for five years. SUMMARY OF PROVISIONS: Section 1: Extends the payment in lieu of taxes for certain property in the Town of Hempstead from June 2023 to June 2028.