Bill S6172 version original - James Skoufis
BILL NUMBER: S6172 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the penal law, in relation to permitting the possession of firearms in sensitive areas for the purposes of performing historical battle reenactments PURPOSE: Permits the possession of firearms in sensitive areas for the purposes of performing historical battle reenactments. SUMMARY OF PROVISIONS: Section 1: amends section 265.