Bill S6211C version original - Monica Martinez
BILL NUMBER: S6211 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the environmental conservation law, in relation to wild or exotic animals PURPOSE: The purpose of this bill is to establish and expand the definition/s of wild and exotic animals to ensure such animals and to prohibit such animals from being imported, sold or owned. SUMMARY OF PROVISIONS: Section 1.