Bill S6194 version original - Kevin S. Parker
BILL NUMBER: S6194 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to requiring instruction in physical education in certain elementary and secondary schools PURPOSE OF BILL: This bill would create mandatory physical education requirements in certain elementary and secondary schools. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Legislative intent. Section 2: Subdivision 5 of section 803 of the education law, as amended by chapter 118 of the laws of 1957, is amended.