Bill S6175 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S6175 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the general municipal law, in relation to the conducting of games of chance by certain organizations PURPOSE: The purpose of this bill is to update the provisions that govern the conduct of games of chance in New York. Article 9-A of the General Municipal Law was first codified in 1967 and has not been substantially modified since that date.