Bill S6186 version original - Robert Rolison
BILL NUMBER: S6186 SPONSOR: ROLISON TITLE OF BILL: An act to amend the penal law, in relation to classifying criminal possession of a firearm from a class E felony to a class D felony PURPOSE OR GENERAL IDEA OF BILL: This bill will change the felony classification of criminal possession of a firearm from a class E felony to a class D felony. SUMMARY OF PROVISIONS: Section 1: Section 265.01-b of the penal law, is amended to read as follows: § 265.01-b Criminal possession of a firearm.