Bill S6227 version original - Brad Hoylman-Sigal
BILL NUMBER: S6227 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to repeal section 260.15 of the penal law relating to affirmative defenses to the charge of endangering the welfare of a child PURPOSE: To repeal subdivision 15 of section 260 of the penal law to remove the affirmative defense against prosecution for endangering the welfare of a child on the basis of religious beliefs or affiliation. SUMMARY OF PROVISIONS: Section 1 repeals section 260.15 of the penal law.