Bill S6221 version original - John Mannion
BILL NUMBER: S6221 SPONSOR: MANNION TITLE OF BILL: An act to amend the penal law, in relation to increasing the penalties for assault on police officers, peace officers, firefighters and emer- gency medical technicians PURPOSE OF BILL: The purpose of this bill is to increase the penalty for criminal acts of assault on law enforcement officers and emergency first responders. SUMMARY OF PROVISIONS: Section 1 of this bill would make technical changes to section 120.05 of the Penal Law.