Bill S6226A version original - Michelle Hinchey
BILL NUMBER: S6226 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the public health law, in relation to availability of ambulance services to store and distribute blood and initiate and admin- ister blood transfusions PURPOSE: The purpose of this bill is to permit ambulance services to store and distribute blood and initiate and administer blood transfusions. SUMMARY OF PROVISIONS: Section 1: amends the public health law to allow for ambulance services to store and distribute blood.