Bill S6234 version original - Andrew J. Lanza
BILL NUMBER: S6234 SPONSOR: LANZA TITLE OF BILL: An act to amend the tax law, in relation to establishing a personal income tax credit for the purchase of a gun safe PURPOSE OR GENERAL IDEA OF BILL: Establishes a personal income tax credit for the purchase of a gun safe SUMMARY OF SPECIFIC PROVISIONS: Section 1. Section 606 of the tax law is amended by adding a new subsection to read as follows. Allowance of credit.