Bill S6239 version original - Kevin Thomas
BILL NUMBER: S6239 SPONSOR: THOMAS TITLE OF BILL: An act to amend the criminal procedure law, in relation to protective orders for juries PURPOSE GENERAL IDEA OF BILL: To ensure the welfare of prospective jurors by allowing their names and addresses to remain anonymous when there is reason to believe that a trial will pose a threat to their safety and wellbeing. SUMMARY OF PROVISIONS: Section 1 amends paragraph (a) of subdivision 1 of section 270.