Bill S6235 version original - Andrew J. Lanza
BILL NUMBER: S6235 SPONSOR: LANZA TITLE OF BILL: An act to amend the election law, in relation to instant run-off voting in the city of New York; and to repeal certain provisions of such law relating thereto PURPOSE: To establish instant run-off voting in primary elections for mayor, public advocate, and comptroller in the city of New York. SUMMARY OF PROVISIONS: Section 1 of the bill repeals Section 6-162 of the election law and adds a new Section 6162.