Bill S6246 version original - Andrew Gounardes
BILL NUMBER: S6246 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to calibration checks for weigh in motion violation monitoring systems PURPOSE OR GENERAL IDEA OF BILL: To amend the vehicle and traffic law in relation to the calibration checks for weight in motion violation monitoring systems SUMMARY OF PROVISIONS: Section 1. Subdivision 3 of section 385-a of the vehicle and traffic law, as added by chapter 773 of the laws of 2021 is amended.