Bill S6266 version original - James Sanders Jr.
BILL NUMBER: S6266 SPONSOR: SANDERS TITLE OF BILL: An act to amend the executive law, in relation to withholding the gover- nor's salary until the legislative passage of the budget occurs PURPOSE OR GENERAL IDEA OF BILL: This bill would withhold the governor's salary until the legislative passage of the budget occurs. SUMMARY OF PROVISIONS: Section 1: 1.