Bill S6289 version original - Jake Ashby
BILL NUMBER: S6289 SPONSOR: ASHBY TITLE OF BILL: An act to incorporate the Salem Volunteer Fire Department Benevolent Association, and providing for its powers and duties PURPOSE: To authorize the formation of a not-for-profit corporation by the name of "Salem Fire Department Benevolent Association" SUMMARY OF PROVISIONS: Section 1: Provides for the incorporation of the association. Section 2: Provides for those eligible for membership in the associ- ation.