Bill S6276 version original - Lea Webb
BILL NUMBER: S6276 SPONSOR: WEBB TITLE OF BILL: An act to amend the tax law, in relation to extending the authorization of the county of Tompkins to impose an additional one percent of sales and compensating use taxes PURPOSE OR GENERAL IDEA OF BILL: This bill authorizes Tompkins County to extend a one percent sales tax for an additional two years. SUMMARY OF PROVISIONS: Section 1: Extends the authorization of Tompkins County to impose an additional one percent sales tax by two years.