Bill S6277 version original - Lea Webb
BILL NUMBER: S6277 SPONSOR: WEBB TITLE OF BILL: An act to amend the tax law, in relation to extending the authorization for imposition of additional sales tax in the county of Cortland PURPOSE OR GENERAL IDEA OF BILL: This bill authorizes Cortland County to extend a one percent sales tax for an additional two years. SUMMARY OF PROVISIONS: Section 1: Extends the authorization of Cortland County to impose an additional one percent sales tax by two years. Section 2: Establishes the effective date.