Bill S6302 version original - Robert Jackson
BILL NUMBER: S6302 SPONSOR: JACKSON TITLE OF BILL: An act to amend the penal law, in relation to creating the offense of operating as a major firearms trafficker PURPOSE: To enhance criminal penalties for major firearm traffickers. SUMMARY OF PROVISIONS: Section 1 creates a new Section 265.46 of the Penal Law that establishes the criminal offense of operating as a major firearms trafficker as an A-1 felony.