Bill S6296 version original - John Mannion
BILL NUMBER: S6296 SPONSOR: MANNION TITLE OF BILL: An act to amend the penal law, in relation to the crimes of falsely reporting an incident in the first and second degrees PURPOSE OF BILL: The purpose of this bill is to expand penalties for false reporting to include the use of a deadly weapon or dangerous instrument. SUMMARY OF PROVISIONS: Section 1 of this bill would amend section 240.