Bill S6322A version original - Robert Jackson
BILL NUMBER: S6322 SPONSOR: JACKSON PURPOSE: This bill grants service credit to members of the uniformed correction force of the NYC Department of Correction covered by provisions of law for over 25 years. SUMMARY OF PROVISIONS: Section 1 amends Section 504 of the Retirement and Social Security Law to add the language "plus one-sixtieth of final average salary times years of credited service in excess of twenty-five". Section 2 is the effective date.