Bill S6317 version original - Michelle Hinchey
BILL NUMBER: S6317 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the public authorities law, in relation to establishing a do-not-solicit solar development statewide registry prohibiting solar solicitors from contacting persons who are included on such registry PURPOSE: To establish a mechanism to prevent the harassment of farmers and farm businesses by Solar Developers. SUMMARY OF PROVISIONS: Section One adds a new section 1885 to the Public Authorities law which: 1.