Bill S6326 version original - Michelle Hinchey
BILL NUMBER: S6326 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the agriculture and markets law, in relation to allowing certain commercial feed to include hemp seed PURPOSE: The purpose of this legislation is to allow for the inclusion of seeds including seed hulls and seed meal from industrial hemp plants in commercial feed solely for pets, specialty pets, horses, and camelids. This legislation would not authorize such inclusion in commercial feed for other livestock.