Bill S6330 version original - Dean Murray
BILL NUMBER: S6330 SPONSOR: MURRAY TITLE OF BILL: An act to amend the penal law, in relation to granting dealers in firearms access to the NYSPIN database PURPOSE: To allow dealers in firearms access to real-time information on the New York State Police Information Network firearm database to increase the chances of catching someone who is trying to sell a stolen firearm. SUMMARY OF PROVISIONS: Section 1 of this bill amends subdivision 2 of section 400.