Bill S6355A version original - Joseph P. Addabbo Jr.
BILL NUMBER: S6355 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the banking law and the state finance law, in relation to credit unions PURPOSE: To create the credit union deposit program. SUMMARY OF PROVISIONS: This bill amends Article 2 and section 454 of the Banking Law and section 98-a and 105 of the State Finance Law to create the credit union deposit program.