Bill S6350B version original - Rachel May
BILL NUMBER: S6350 SPONSOR: MAY TITLE OF BILL: An act to amend the education law, in relation to enacting the "freedom to read act" PURPOSE: To ensure school libraries and libraries are able to offer a wide age- appropriate materials to students. SUMMARY OF PROVISIONS: Adds a new subdivision to Article 2 Section 13 of the Education Law to ensure school libraries and librarians are able to provide students with a wide array of age-appropriate materials in schools.