Bill S3473A version A - Toby Ann Stavisky
BILL NUMBER: S3473A SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to trustees of the state university and city university PURPOSE OR GENERAL IDEA OF BILL: This bill ensures that Gubernatorial and Mayoral appointments to the State University and City University Boards of Trustees are independent of the appointing entities. SUMMARY OF SPECIFIC PROVISIONS: Section 1.