Bill S6349A version original - Neil D. Breslin
BILL NUMBER: S6349 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the executive law, in relation to adding Chuctanunda creek to the definition of inland waterways PURPOSE OR GENERAL IDEA OF BILL: Adding Chuctanunda Creek to the list of New York State's major creeks. SUMMARY OF PROVISIONS: Subdivision 4 of section 911 of the executive law is amended to add Chuctanunda Creek to the list of New York State's major creeks.